Scripts conflict fix
Apparently another script is adding a Font Awesome into the text of the element that holds the nickname, used on the API request, this fix should hopefully help with getting just the nickname, but it's not a long-term solution. Also added a custom header with the requests so I can keep better track of it.